requirements.txt
asgiref==3.2.3
Django==3.0
django-cors-headers==3.2.0
django-environ==0.4.5
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
Pillow==6.2.1
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
